:technologist: Organize OPDS into versioned modules
This is mostly in preparation for supporting v2.0
